rag-lab/requirements.txt
openai>=1.0
sentence-transformers>=2.2
faiss-cpu>=1.7
numpy>=1.24
python-dotenv>=1.0
jupyter>=1.0
